+++ CHANGELOG.md
@@ -5,6 +5,9 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 
 ## [Unreleased]
 
+### Fixed
+- Handle ACCEPTED(202) response type set in raml for asynchronous REST endpoints
+
 ## [1.2.0] - 2017-01-19
 
 ### Added